remove deleted files
